remove old commercial tag page
