--- admin/app/views/commercialReports/targets.scala.html
@@ -1,27 +1,0 @@
-@(env: String, sponsoredTags: Seq[dfp.Sponsorship], advertisementTags: Seq[dfp.Sponsorship], pageSkinnedAdUnits: Seq[String])
-@import conf.AdminConfiguration.pa
-@import org.joda.time.{DateTime, DateMidnight}
-
-@admin_main("Commercial reporter", env, isAuthed = true) {
-
-<h1>Sponsored Tags</h1>
-<ol>
-@for(tag <- sponsoredTags) {
-    <li>@{tag.tags.mkString(", ")}@{tag.sponsor.map(sponsor => <small>[sponsored by <b>@sponsor</b>]</small>)}</li>
-}
-</ol>
-
-<h1>Advertisement Feature Tags</h1>
-<ol>
-@for(tag <- advertisementTags) {
-    <li>@{tag.tags.mkString(", ")}@{tag.sponsor.map(sponsor => <small>[sponsored by <b>@sponsor</b>]</small>)}</li>
-}
-</ol>
-
-<h1>Pageskinned Ad Units</h1>
-<ol>
-    @for(adunit <- pageSkinnedAdUnits) {
-    <li>@adunit</li>
-    }
-</ol>
-}
