+++ admin/public/css/commercial.css
@@ -2,6 +2,16 @@
     margin: 0 auto 20px;
 }
 
+.render-time {
+    background-repeat: no-repeat;
+    background-position: 50% 50%;
+    margin-bottom: 30px;
+    min-height: 200px;
+}
+.render-time--ad {
+    min-height: 235px;
+}
+
 .render-time--ad__form {
     padding-left: 55px;
     margin-bottom: 0;
@@ -13,11 +23,4 @@
 .render-time--ad__form__select {
     margin-bottom: 0;
     margin-left: 10px;
-}
-
-.render-time {
-    background-repeat: no-repeat;
-    background-position: 50% 50%;
-    margin-bottom: 30px;
-    min-height: 200px;
 }
